+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 *Also see [Tools and installer changelog](https://github.com/clojure/brew-install/blob/1.10.3/CHANGELOG.md)*
 
 * next
+  * Remove deprecated keys from basis
   * Update to tools.deps.cli 0.9.21
 * 0.17.1297 on Mar 3, 2023
   * Add new function - resolve-added-libs